start moving the cli to examples
